Update the description text and remove the ProblemInputDisplay file.
